LayoutTests/fast/css/pseudo-indeterminate-with-radio-buttons-style-invalidation-expected.txt

Verify that we do not invalidate more than needed to satisfy :indeterminate

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS elementsNeedingStyleRecalc() is []
PASS elementsWithIndeterminateStyle() is ["radio1", "radio2", "radio3", "radio4", "radio5", "radio6", "radio7", "radio8", "radio9", "radio10"]
PASS checkedElements() is []
Check radio3
PASS elementsNeedingStyleRecalc() is ["radio1", "radio2", "radio3", "radio4", "radio5"]
PASS elementsWithIndeterminateStyle() is ["radio6", "radio7", "radio8", "radio9", "radio10"]
PASS checkedElements() is ["radio3"]
Check radio8
PASS elementsNeedingStyleRecalc() is ["radio6", "radio7", "radio8", "radio9", "radio10"]
PASS elementsWithIndeterminateStyle() is []
PASS checkedElements() is ["radio3", "radio8"]
Check radio4
PASS elementsNeedingStyleRecalc() is ["radio3", "radio4"]
PASS elementsWithIndeterminateStyle() is []
PASS checkedElements() is ["radio4", "radio8"]
Check radio9
PASS elementsNeedingStyleRecalc() is ["radio8", "radio9"]
PASS elementsWithIndeterminateStyle() is []
PASS checkedElements() is ["radio4", "radio9"]
PASS successfullyParsed is true

TEST COMPLETE
